// src/services/Reline/interfaces/edge.relineSchema.ts
import { RelineElementBaseSchema } from "./elementBase.relineSchema";
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface RelineEdgeSchema extends RelineElementBaseSchema{
  id: string;
  source: string;
  target: string;
  metadata: {
    basic: {
      id: string;
      label?: string;
      [k: string]: unknown;
    };
    [k: string]: {
      [k: string]: unknown;
    };
  };
}